Update 3 files
- /kpm_e2sm/src/kpm/encode_kpm_indication.cpp
- /kpm_e2sm/src/kpm/encode_kpm_indication.hpp
- /kpm_e2sm/src/kpm/kpm_callbacks.cpp

+++ kpm_e2sm/src/kpm/kpm_callbacks.cpp
@@ -36,7 +36,7 @@ using namespace std;
 struct KPIEntry {
   std::string timestamp;
   double rx_brate_uplink_Mbps;
-  int ul_n_samples;
+  unsigned long ul_n_samples;
 };
 
 int create_sctp_socket(const std::string& ip, int port) {
